+++ core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 1.2.1 (7. April 2020)
+
++ [#2008](https://github.com/luyadev/luya/pull/2008) Default values for JsonBehavior was wrong declared, expected is the opposite way 1. array as validation rule, 2. auto decode data after find.
+
 ## 1.2.0 (7. April 2020)
 
 + [#2003](https://github.com/luyadev/luya/pull/2003) Add new `StringHelper::isNumeric()` which checks whether a value is nummeric (with regex instead of `is_numeric`) or not.